Conflicts:
	test/unit/common.h
	tpie/CMakeLists.txt

+++ tpie/util.h
@@ -73,7 +73,7 @@ struct linear_memory_base {
 	///////////////////////////////////////////////////////////////////////////
 	inline static stream_size_type memory_usage(stream_size_type size) {
 		return static_cast<stream_size_type>(
-			floor(size * child_t::memory_coefficient() + child_t::memory_overhead()));
+			floor(static_cast<double>(size) * child_t::memory_coefficient() + child_t::memory_overhead()));
 	}
 
 	///////////////////////////////////////////////////////////////////////////
@@ -88,6 +88,10 @@ struct linear_memory_base {
 	}
 };
 
+///////////////////////////////////////////////////////////////////////////////
+/// \brief Computes the least integer strictly greater than log(t). Maybe this
+/// ought to compute something different. Used in array.h.
+///////////////////////////////////////////////////////////////////////////////
 template <int t>
 struct template_log {
 	static const size_t v=1+template_log< t/2 >::v;
